chore: cached fetch token price
